requirements.txt
flask
google-generativeai
python-dotenv
Pillow
pandas
pydantic
openpyxl